+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Unreleased
 
-- [added] `messaging.AndroidNotification`type now supports channel_id.
+- [added] `messaging.Aps` type now supports configuring a critical sound.
+  A new `messaging.CriticalSound` type has been introduced for this purpose.
+- [added] `messaging.AndroidNotification` type now supports `channel_id`.
 
 # v6.3.0